code.vision
This is where you code all of your code. Use the OnConsole: <anything you want to put here> command to put text on the console. Use the SystemPause: <number of seconds> command to stop the program for a number of seconds. Use the OnGuiController: <title> command to create a gui interface. Have fun!